refactoring of jdbc tests
