+++ infobip-spring-data-jdbc-querydsl/src/test/java/com/infobip/spring/data/jdbc/PersonRepository.java
@@ -1,0 +1,4 @@
+package com.infobip.spring.data.jdbc;
+
+public interface PersonRepository extends QuerydslJdbcRepository<Person, QPerson, Long> {
+}
