+++ infobip-spring-data-jdbc-querydsl/src/test/java/com/infobip/spring/data/jdbc/PersonSettingsRepository.java
@@ -1,0 +1,5 @@
+package com.infobip.spring.data.jdbc;
+
+public interface PersonSettingsRepository extends QuerydslJdbcRepository<PersonSettings, QPersonSettings, Long> {
+
+}
