--- infobip-spring-data-jdbc-querydsl/src/test/java/com/infobip/spring/data/jdbc/person/PersonRepository.java
@@ -1,7 +1,0 @@
-package com.infobip.spring.data.jdbc.person;
-
-import com.infobip.spring.data.jdbc.QPerson;
-import com.infobip.spring.data.jdbc.QuerydslJdbcRepository;
-
-public interface PersonRepository extends QuerydslJdbcRepository<Person, QPerson, Long> {
-}
