--- infobip-spring-data-jdbc-querydsl/src/test/java/com/infobip/spring/data/jdbc/person/settings/PersonSettingsRepository.java
@@ -1,8 +1,0 @@
-package com.infobip.spring.data.jdbc.person.settings;
-
-import com.infobip.spring.data.jdbc.QPersonSettings;
-import com.infobip.spring.data.jdbc.QuerydslJdbcRepository;
-
-public interface PersonSettingsRepository extends QuerydslJdbcRepository<PersonSettings, QPersonSettings, Long> {
-
-}
